// interfaces/clone-checklist.ts
export interface ICloneChecklistServiceLine {
  serviceLineKey: string
  taskId: number
  taskName?: string
  serviceId: number
  chargeName?: string
  statusName?: string
  quantity: number
  amount: number
  referenceNo?: string
  remarks?: string
  bargeName?: string
  crewName?: string
  visaTypeName?: string
  awbNo?: string
}